ajuste: botao de recalcular na tela de periodos parciais esta removendo lancamentos e estudo ja liberado.. agora dara a msg de existe estudo para o perido xx.. excluir o estudo e tentar novamente

+++ nds-service-client-impl/src/main/java/br/com/abril/nds/service/impl/ParciaisServiceImpl.java
@@ -489,7 +489,10 @@ public class ParciaisServiceImpl implements ParciaisService{
 					
 					Lancamento lancamento = item.getLancamentoPeriodoParcial();
 					
-					if(lancamento!= null && getStatusLancamentoPreExpedicao().contains(lancamento.getStatus())){
+					if(lancamento!= null && getStatusLancamentoPreExpedicao().contains(lancamento.getStatus())) {
+					  if (  lancamento.getEstudo() != null  ){
+						  throw new ValidacaoException(TipoMensagem.WARNING,"Existe estudo para o período "+item.getNumeroPeriodo()+".Exclua o estudo e tente novamente.");
+					  }
 						
 						periodoLancamentoParcialRepository.remover(item);
